addressed review comments: git rm inl files, removing unnecessary headers, move simulator.h code
