--- systems/analysis/antiderivative_function-inl.h
@@ -1,5 +1,0 @@
-#pragma once
-
-// Instantiation of AntiderivateFunction<T> for T other than double
-// requires ScalarInitialValueProblem<T> definition.
-#include "drake/systems/analysis/scalar_initial_value_problem-inl.h"
