+++ systems/analysis/antiderivative_function.cc
@@ -1,13 +1,5 @@
 #include "drake/systems/analysis/antiderivative_function.h"
 
-#include "drake/common/default_scalars.h"
-#include "drake/systems/analysis/scalar_initial_value_problem.h"
-
-namespace drake {
-namespace systems {
-
 DRAKE_DEFINE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_NONSYMBOLIC_SCALARS(
-    class AntiderivativeFunction);
+    class drake::systems::AntiderivativeFunction)
 
-}  // namespace systems
-}  // namespace drake
