+++ systems/analysis/implicit_integrator.cc
@@ -3,7 +3,6 @@
 #include <Eigen/LU>
 
 #include "drake/common/autodiff.h"
-#include "drake/common/default_scalars.h"
 #include "drake/math/autodiff_gradient.h"
 
 namespace drake {
@@ -428,9 +427,9 @@ bool ImplicitIntegrator<T>::MaybeFreshenMatrices(
   }
 }
 
-DRAKE_DEFINE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_NONSYMBOLIC_SCALARS(
-    class ImplicitIntegrator)
-
 }  // namespace systems
 }  // namespace drake
+
+DRAKE_DEFINE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_NONSYMBOLIC_SCALARS(
+    class drake::systems::ImplicitIntegrator)
