+++ systems/analysis/initial_value_problem.h
@@ -1,9 +1,7 @@
 #pragma once
 
 #include <memory>
-#include <optional>
 #include <utility>
-#include <vector>
 
 #include "drake/common/default_scalars.h"
 #include "drake/common/drake_copyable.h"
@@ -11,8 +9,6 @@
 #include "drake/systems/analysis/dense_output.h"
 #include "drake/systems/analysis/integrator_base.h"
 #include "drake/systems/framework/context.h"
-#include "drake/systems/framework/parameters.h"
-#include "drake/systems/framework/vector_base.h"
 
 namespace drake {
 namespace systems {
@@ -263,8 +259,8 @@ class InitialValueProblem {
   std::unique_ptr<IntegratorBase<T>> integrator_;
 };
 
-DRAKE_DECLARE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_NONSYMBOLIC_SCALARS(
-    class InitialValueProblem)
-
 }  // namespace systems
 }  // namespace drake
+
+DRAKE_DECLARE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_NONSYMBOLIC_SCALARS(
+    class drake::systems::InitialValueProblem)
