+++ systems/analysis/integrator_base.cc
@@ -1,7 +1,5 @@
 #include "drake/systems/analysis/integrator_base.h"
 
-#include "drake/common/default_scalars.h"
-
 namespace drake {
 namespace systems {
 
@@ -462,9 +460,9 @@ typename IntegratorBase<T>::StepResult
   }
 }
 
-DRAKE_DEFINE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_NONSYMBOLIC_SCALARS(
-    class IntegratorBase)
-
 }  // namespace systems
 }  // namespace drake
+
+DRAKE_DEFINE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_NONSYMBOLIC_SCALARS(
+    class drake::systems::IntegratorBase)
