+++ systems/analysis/lyapunov.h
@@ -1,8 +1,7 @@
 #pragma once
 
-#include <functional>
-
-#include "drake/common/symbolic.h"
+#include "drake/common/autodiff.h"
+#include "drake/common/eigen_types.h"
 #include "drake/systems/framework/context.h"
 #include "drake/systems/framework/system.h"
