+++ systems/analysis/scalar_initial_value_problem.cc
@@ -1,12 +1,4 @@
 #include "drake/systems/analysis/scalar_initial_value_problem.h"
 
-#include "drake/common/default_scalars.h"
-
-namespace drake {
-namespace systems {
-
 DRAKE_DEFINE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_NONSYMBOLIC_SCALARS(
-    class ScalarInitialValueProblem)
-
-}  // namespace systems
-}  // namespace drake
+    class drake::systems::ScalarInitialValueProblem)
